refactor: address code review for reference resolution
- Extract reference fetch into Downloader.referenceFetch so the download task and
  integration test share one mapping (test now guards production code, not a copy)
- Dedup references within a single schema (.distinct) — a schema declaring the same
  (subject, version) twice no longer triggers a redundant fetch; add a unit test
- Document the fetch-echoes-subject precondition on ReferenceResolver.resolve
- Correct the "shared cache avoids double fetch" claim in plan/research/contracts:
  resolution uses getSchemaMetadata (not cached vs getByVersion), so it adds ~1
  metadata round-trip per schema — accepted, graphs are small
